# perf(db): RawTable, decode/encode only if needed