--- a/crates/storage/db/src/tables/mod.rs
+++ b/crates/storage/db/src/tables/mod.rs
@@ -14,7 +14,10 @@
 
 pub mod codecs;
 pub mod models;
+mod raw;
 pub(crate) mod utils;
+
+pub use raw::{RawDubSort, RawKey, RawTable, RawValue};
 
 /// Declaration of all Database tables.
 use crate::{