Got around ~10% more performance when moving decode/encode compress/decompress to parallel threads by creating the `RawTable`, `RawKey` and `RawValue` it allows us to in nice way directly decide what we want to do with data.

```
    Finished bench [optimized] target(s) in 0.22s
     Running benches/criterion.rs (target/release/deps/criterion-f89f26f1148427ff)
Stages/AccountHashing   time:   [27.190 ms 27.438 ms 27.770 ms]
                        change: [-2.4274% -0.7469% +0.8175%] (p = 0.42 > 0.05)
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
```

Measurement from previous PR: https://github.com/paradigmxyz/reth/pull/2077

```
    Finished bench [optimized] target(s) in 5.10s
     Running benches/criterion.rs (target/release/deps/criterion-f89f26f1148427ff)
Stages/AccountHashing   time:   [31.141 ms 31.366 ms 31.668 ms]
                        change: [-52.099% -51.767% -51.443%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
  ```
  
From 31.3ms to 27.4ms.

copilot:all