@gakonst 2023-04-03
@joshieDo @onbjerg is this idea of RawKey something we'd want to introduce in other places in the stage/db accesses? it looks like we can get some perf by lazily decoding when reading from the db?