@joshieDo 2023-04-03
I wonder what I've done wrong, because when I tried something of the sort... I didn't have a significant improvement

it's good for any insertion/reading that deals mostly with fixed sized keys and values