@rakita 2023-04-03
> @joshieDo @onbjerg is this idea of RawKey something we'd want to introduce in other places in the stage/db accesses? it looks like we can get some perf by lazily decoding when reading from the db? seems like we could also apply this to SendersRecovery?

This can be useful to implement Header views as in get header compressed data, don't decompress it but just read needed values is in for example "timestamp". And as @joshieDo said everywhere we have parallelization (rayon) we can use it.

@gakonst SenderRecovery I tried [here](https://github.com/paradigmxyz/reth/pull/2086) and didn't get improvements, that stage is very IO bound as we have parallelized recovery to the max. Every stage that reads all tx is going to be IO bound as it contains a lot of entries. 


> I wonder what I've done wrong, because when I tried something of the sort... I didn't have a significant improvement
> 
> Wherever stage we're using rayon to handle data, we could move the compression/encoding there with this if necessary. I think the main benefit here comes from skipping the decompress & compress of `account` which is unnecessary. Key should be actually optimized by rust anyway since it's a fixed size array?

Probably, it is relevant mostly for compress/decompress of big stractures and in this case as you said it skipped that part for `account`. I don't know if would be optimized, as transfer of data is happening for `H` types it would move from heap to stack but in essence, this is an insignificant penalty.